+++ src/core/index.ts
@@ -1,0 +1,4 @@
+import './_normalize.scss';
+import './_base.scss';
+import './_css_var.scss';
+import './_colors.scss';